+++ CHANGELOG.md
@@ -5,7 +5,7 @@ so user could call `String.red((ls))` and got the expected result.
 - add `grep` for String module.
 - add `highlight` for Parse module.
 - pull down echo for function declare.
-- allow var after `!` like `!$x`.
+- allow var after `!`/`=` like `!$x`, `=$x`
 - add `quit` to repl.
 while `exit [status]` used for script, `quit` used for repl.