+++ CHANGELOG.md
@@ -6,7 +6,7 @@ so user could call `String.red((ls))` and got the expected result.
 - add `highlight` for Parse module.
 - pull down echo for function declare.
 - allow var after `!`/`=` like `!$x`, `=$x`
-- add `quit` to repl.
+- add `exit` to repl.
 while `exit [status]` used for script, `quit` used for repl.
 - expand home during `for`,`use` and in `Fs` functions.